requirements.txt
pandas
numpy<2.0.0
matplotlib
mplfinance
torch
torchvision
scikit-learn
backtest-pro
Pillow
PyYAML
torchinfo
tensorboard
torchmetrics
